+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+* Could not use `loadPipelineScriptForTest(...)` with Groovy scripts whose filenames were not valid Java class names.
+
 ## 2.1.0
 
 _Release Date: 2020-02-04_